alvium-fmc-zcu106: update to Vivado 2025.1 based fpga design
Update the fpga designs to a Vivado 2025.1 based versions. The new
designs support a higher image data throughput by transferring 4 pixels
per clock in the AXI streams between the CSI receiver and the video
framebuffer write. This is required, because starting with firmware 14
the speed of the Alvium C series cameras is increased, which caused
buffer overflows in the old designs. The new designs also use the
S_AXI_HPC instead of the S_AXI_HP ports of the ZynqMP.